fix: only set script hashes in manifest V2 CSP
chrome will refuse to load manifest V3 extensions containing script
hashes in CSP
